+++ R/downloads.R
@@ -1,19 +1,21 @@
 #' Prepare spectrum download
 #'
-#' @param hintr_output object
+#' @param output hintr output object
 #' @param path Path to save output file
+#' @param notes User added notes from front end of app as a string
 #'
 #' @return Path to output file and metadata for file
 #' @export
 hintr_prepare_spectrum_download <- function(output,
-                                            path = tempfile(fileext = ".zip")) {
+                                            path = tempfile(fileext = ".zip"),
+                                            notes = NULL) {
   assert_model_output_version(output)
   progress <- new_simple_progress()
   progress$update_progress("PROGRESS_DOWNLOAD_SPECTRUM")
   model_output <- readRDS(output$model_output_path)
   options <- yaml::read_yaml(text = model_output$info$options.yml)
   list(
-    path = save_output_spectrum(path, model_output$output_package),
+    path = save_output_spectrum(path, model_output$output_package, notes),
     metadata = list(
       description = build_output_description(options),
       areas = options$area_scope,
@@ -24,7 +26,7 @@ hintr_prepare_spectrum_download <- function(output,
 
 #' Prepare coarse age group download
 #'
-#' @param hintr_output object
+#' @param output hintr output object
 #' @param path Path to save output file
 #'
 #' @return Path to output file and metadata for file
@@ -47,7 +49,7 @@ hintr_prepare_coarse_age_group_download <- function(
 
 #' Prepare summary report download
 #'
-#' @param hintr_output object
+#' @param output hintr output object
 #' @param path Path to save output file
 #'
 #' @return Path to output file and metadata for file
@@ -66,6 +68,33 @@ hintr_prepare_summary_report_download <- function(
       description = build_summary_report_description(options),
       areas = options$area_scope,
       type = "summary"
+    )
+  )
+}
+
+#' Prepare comparison report download
+#'
+#' @param hintr_output object
+#' @param path Path to save output file
+#'
+#' @return Path to output file and metadata for file
+#' @export
+hintr_prepare_comparison_report_download <- function(
+  output, path = tempfile(fileext = ".html")) {
+  assert_model_output_version(output)
+  progress <- new_simple_progress()
+  progress$update_progress("PROGRESS_DOWNLOAD_COMPARISON")
+  model_output <- readRDS(output$model_output_path)
+  options <- yaml::read_yaml(text = model_output$info$options.yml)
+  list(
+    ## Just using the summary report as a placeholder - eventually
+    ## this will call a separate method to generate comparison report
+    path = generate_output_summary_report(path, output$model_output_path,
+                                          quiet = TRUE),
+    metadata = list(
+      description = build_comparison_report_description(options),
+      areas = options$area_scope,
+      type = "comparison"
     )
   )
 }